move playassetvideowithrage to api client
